multiple changes
Problems solved this weekend
Prevent background scrolling when hamburger menu is open
darken the background color
Invert hamburger menu color
Deprecated Velocity Plugin
Loading actual time applied

Opinion
If it rains, the letters will not be visible.
- The text and images suddenly became large, but I set the magnifying glass to 90%.
I don't know if it was because of that, but I'll reduce the size.